Revert structural changes
