+++ .gitignore
@@ -18,6 +18,7 @@
 *.swp
 
 # Generated UniFFI artifacts (regenerated in CI/local scripts)
+/target/uniffi/
 /bindings/swift/
 /bindings/kotlin/
 /bindings/kotlin-android/org/
