+++ scripts/ci/build_android_jni.sh
@@ -4,9 +4,10 @@ set -euo pipefail
 ROOT_DIR="$(cd "$(dirname "${BASH_SOURCE[0]}")/../.." && pwd)"
 cd "$ROOT_DIR"
 
-rm -rf bindings/kotlin-android/jniLibs
-cargo ndk -t arm64-v8a -o bindings/kotlin-android/jniLibs build --release --features uniffi --lib
-cargo ndk -t armeabi-v7a -o bindings/kotlin-android/jniLibs build --release --features uniffi --lib
-cargo ndk -t x86_64 -o bindings/kotlin-android/jniLibs build --release --features uniffi --lib
+OUT_DIR="target/uniffi/kotlin-android/jniLibs"
+rm -rf "$OUT_DIR"
+cargo ndk -t arm64-v8a -o "$OUT_DIR" build --release --features uniffi --lib
+cargo ndk -t armeabi-v7a -o "$OUT_DIR" build --release --features uniffi --lib
+cargo ndk -t x86_64 -o "$OUT_DIR" build --release --features uniffi --lib
 
-echo "Built Android JNI libs in bindings/kotlin-android/jniLibs"
+echo "Built Android JNI libs in $OUT_DIR"
