+++ scripts/ci/uniffi_generate_swift.sh
@@ -4,10 +4,13 @@ set -euo pipefail
 ROOT_DIR="$(cd "$(dirname "${BASH_SOURCE[0]}")/../.." && pwd)"
 cd "$ROOT_DIR"
 
+OUT_DIR="target/uniffi/swift"
+mkdir -p "$OUT_DIR"
+
 cargo run --manifest-path bindings/uniffi-bindgen/Cargo.toml -- \
   generate bindings/rgb_lightning_node.udl \
   --language swift \
   --config uniffi.toml \
-  -o bindings/swift
+  -o "$OUT_DIR"
 
-echo "Generated Swift bindings in bindings/swift"
+echo "Generated Swift bindings in $OUT_DIR"
