--- src/routes/payments/mod.rs
@@ -1,12 +1,0 @@
-use super::*;
-
-mod read;
-mod write_payments;
-mod write_swaps;
-
-pub(crate) use read::{
-    decode_ln_invoice, decode_rgb_invoice, get_payment, get_swap, invoice_status, list_payments,
-    list_swaps,
-};
-pub(crate) use write_payments::{keysend, ln_invoice, send_payment};
-pub(crate) use write_swaps::{maker_execute, maker_init, taker};
