+++ src/sdk/payments.rs
@@ -25,7 +25,7 @@ pub(crate) async fn list_payments(state: Arc<AppState>) -> Result<Vec<PaymentDat
             asset_id,
             payment_hash: hex_str(&payment_hash.0),
             inbound: true,
-            status: payment_info.status.into(),
+            status: payment_info.status,
             created_at: payment_info.created_at,
             updated_at: payment_info.updated_at,
             payee_pubkey: payment_info.payee_pubkey.to_string(),
@@ -51,7 +51,7 @@ pub(crate) async fn list_payments(state: Arc<AppState>) -> Result<Vec<PaymentDat
             asset_id,
             payment_hash: hex_str(&payment_hash.0),
             inbound: false,
-            status: payment_info.status.into(),
+            status: payment_info.status,
             created_at: payment_info.created_at,
             updated_at: payment_info.updated_at,
             payee_pubkey: payment_info.payee_pubkey.to_string(),
@@ -95,7 +95,7 @@ pub(crate) async fn get_payment(
                 asset_id,
                 payment_hash: hex_str(&payment_hash.0),
                 inbound: true,
-                status: payment_info.status.into(),
+                status: payment_info.status,
                 created_at: payment_info.created_at,
                 updated_at: payment_info.updated_at,
                 payee_pubkey: payment_info.payee_pubkey.to_string(),
@@ -122,7 +122,7 @@ pub(crate) async fn get_payment(
                 asset_id,
                 payment_hash: hex_str(&payment_hash.0),
                 inbound: false,
-                status: payment_info.status.into(),
+                status: payment_info.status,
                 created_at: payment_info.created_at,
                 updated_at: payment_info.updated_at,
                 payee_pubkey: payment_info.payee_pubkey.to_string(),
